Fix 'Add task' button bug + add default projects
After clicking on the 'add task' button, the block stays there and each time the button appears it creates a new one with the same name. Now each time that block is deleted.
The user will get some default projects and tasks if he visits the page for the first time.